% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_exiftool}
\alias{check_exiftool}
\title{A simple check for exiftool}
\usage{
check_exiftool(toolname = "exiftool")
}
\description{
Ideally this is not breaking, i.e. does not cause an error immediately. The behaviour of ggsave should default to ggplot2::ggsave if exiftool is not available, because we don't want to disturb the workflow unncessasarily.
}
