% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{ggsave}
\alias{ggsave}
\title{Save a PDF plot along with the name of the file used to create it.}
\usage{
ggsave(filename, plot, cleanup = T, scriptname = NULL, ...)
}
\arguments{
\item{filename}{Destination to save the plot. Must be a PDF.}

\item{plot}{GGplot object to save.}

\item{cleanup}{Remove backup of PDF generated by ExifTool}

\item{scriptname}{Optional override to the currently open Rstudio file, or for use in the case that Rstudioapi is not available.}

\item{...}{Other arguments which are passed directly to ggplot2::ggsave}
}
\description{
Save a PDF plot along with the name of the file used to create it.
}
