% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{ggsource}
\alias{ggsource}
\title{Recover the source code used to create a plot.}
\usage{
ggsource(filename)
}
\arguments{
\item{filename}{Path to PDF plot generated with ggsource::ggsave.}
}
\value{
Either opens the file in Rstudio or prints the path.A
}
\description{
Recover the source code used to create a plot.
}
